Move ReviewVotingPage to review components directory
